Port derive function to C++
Port the derive function from Python to C++. Some trickery is needed,
since the FFT function provided by Eigen doesn't handle matrices, only
vectors. Some extra copying is done, but most of the need for that goes
away when Python is out of the picture.

It's important that FFTW or the intel libs are used, because the default
Eigen FFT is super slow.