Store system in skyline format
Store the system in a memory efficient skyline format. The system
consists of diagonal and symmetric blocks and can thus be stored very
efficiently by only storing the uppermost non-zero diagonals. The
matrix-vector product can not be solved directly by eigen when stored in
this, therefore a matrix-free solver is used.