Build binaries on centos-6 on travis
Build and test the binaries in a CentOS6/RHEL6 image. This is the target
runtime environment, and building on the stock travis ubuntu image
causes conflicts with libc minimum versions.

Unfortunately, that means features like cppcheck and scan-build are more
clunky and will have to wait. They might be re-introduced at some point,
but the test suite will until then be ran inside the CentOS image.