Update Python tests
Simplify the Python tests using comparison functionality provided by
NumPy. We emply tolerances that are stricter than the default parameters
used in `diff.py`. This also replaces tests that are questionable in
`diff.py`, e.g., the average deviation is compared to a positive
constant, but the calculation of the average deviation allows for
negative.