Fixing tests again
We don't need to test for inline vs. crossline sorting from the Python
side. Data is availabe in a NumPy array so independent of which of the
axes is the inline and which one is the crossline axes, access time is
given by the memory layout. This means that for data[i, j, k], i is the
slowest and k is the fastest index.