Restructure CMake config for fine-grained build
We introduce additional build options to steer which components of si4ti
are build. This can be useful if one does not need all components of
si4ti, for example, in the upcoming Python bindings, as it allows us to
reduce the amount of dependencies that need to be installed, may reduces
build time etc. By default, the same amount of command line tools are
build as before. Additionally, we make sure to enable dependencies more
specific to the components being build.

- Add option to only build impedance or timeshift tool.
- Loosen dependency on FFTW. Now, FFTW only needs to be present if the
  flag `USE_FFTW` is set to `ON` (or an equivalent value).
- Boost is only needed if timeshift code is being build.