Rearrange impedance code to be more general
This is a preparatory step for the Python bindings. It moves all of the
actual logic of computing the impedance, including processing the input
and output data into the header file. It will allow us to separate
between pure processing and computing (in the header) and handling of
input/output specific information, like file format, in the functions
calling the impedance computation (Python bindings or command line
tool).