Enable vectorisation for prebuilt wheels
Vectorisation is crucial for good performance in si4ti as Eigen relies
heavily on vectorisation. In tests on a x86_64 platform (Intel Xeon
Platinum 8462Y+ using 8 cores), enabling target specific optimisation
`x86-64-v3` [1] showed a reduction of runtime by ~40% (from 60s to 38s
wallclock time). Enabling newer archtitectures or `-march=native` did
not show improved performance during our tests and would restrict the
portability of the build more severely.

This impacts usability of the wheels as they only work on devices with
support for the used instruction set. This should still be true for all
reasonably new CPUs as as `x86-64-v3` refers to common CPUs with AVX2
such as Intel Haswell (released 2013) and AMD Excavator (released 2015)
and newer "big cores" [2]. It is expected that people running the si4ti
application are on powerful, reasonably new machines, and appreciate the
performance gains. Mac users on arm64 Macs should be unaffected since we
do not turn any vectorisation features on besides the ones that are
activated by default.

[1]: https://gcc.gnu.org/onlinedocs/gcc/x86-Options.html
[2]: https://en.wikipedia.org/wiki/X86-64#Microarchitecture_levels