Add publication workflow
The publication workflow is based on cibuildhweel [1] and is based on
the workflow of other repositories such as segyio [2]. The workflow is
pre-approved to run to publish as trusted publisher [3].

We currently only build wheels that are relevant for the foreseen
applications (Linux) and other platforms for which creation of wheels
was rather straightforward. Therefore, no Windows wheels are built.

The source distribution (sdist) is not build as it would require some
larger restructuring of the repository. It appears to be non-trivial to
package the sdist if the `pyproject.toml` is not in the root directory
of all files to be included in the sdist. In our case, the impedance
source code, test data, README etc. are all accessible from the parent
directory of `pyproject.toml` instead of the directory the
`pyproject.toml` resides in.

[1]: https://cibuildwheel.pypa.io/en/stable/
[2]: https://github.com/equinor/segyio
[3]: https://docs.pypi.org/trusted-publishers/using-a-publisher/