Bump project version number
The previous version, 1.1.0a1, was released manually and is only
available on PyPI as it was inteded for testing the Python bindings. We
bump the version number to 1.1.0a2 with the intend to tag it. This
should trigger the publication workflow to push new wheels to PyPI and
lets us create a new release on GitHub.